Moves TopologyArgument to spec
